+++ packages/react-liveness/src/components/FaceLivenessDetector/shared/FaceLivenessErrorModal.tsx
@@ -8,6 +8,7 @@ import { LivenessErrorState } from '../service';
 import { Toast } from './Toast';
 import { Overlay } from './Overlay';
 import { defaultErrorDisplayText, ErrorDisplayText } from '../displayText';
+import { LivenessClassNames } from '../types/classNames';
 
 export interface CheckScreenComponents {
   ErrorView?: React.ComponentType<FaceLivenessErrorModalProps>;
@@ -30,6 +31,8 @@ const renderToastErrorModal = (props: {
     timeoutMessageText,
     faceDistanceHeaderText,
     faceDistanceMessageText,
+    multipleFacesHeaderText,
+    multipleFacesMessageText,
     clientHeaderText,
     clientMessageText,
     serverHeaderText,
@@ -48,6 +51,10 @@ const renderToastErrorModal = (props: {
       heading = faceDistanceHeaderText;
       message = faceDistanceMessageText;
       break;
+    case LivenessErrorState.MULTIPLE_FACES_ERROR:
+      heading = multipleFacesHeaderText;
+      message = multipleFacesMessageText;
+      break;
     case LivenessErrorState.RUNTIME_ERROR:
       heading = clientHeaderText;
       message = clientMessageText;
@@ -60,14 +67,9 @@ const renderToastErrorModal = (props: {
 
   return (
     <>
-      <Flex
-        gap="xs"
-        alignItems="center"
-        justifyContent="center"
-        color="font.error"
-      >
+      <Flex className={LivenessClassNames.ErrorModal}>
         <AlertIcon ariaHidden variation="error" />
-        <Text fontWeight="bold">{heading}</Text>
+        <Text className={LivenessClassNames.ErrorModalHeading}>{heading}</Text>
       </Flex>
       {message}
     </>
@@ -113,7 +115,7 @@ export const FaceLivenessErrorModal: React.FC<FaceLivenessErrorModalProps> = (
   const { tryAgainText } = displayText;
 
   return (
-    <Overlay backgroundColor="overlay.40">
+    <Overlay className={LivenessClassNames.OpaqueOverlay}>
       <Toast>
         {children}
         <Flex justifyContent="center">