+++ packages/react-storage/src/components/StorageImage/StorageImage.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import classNames from 'classnames';
 
 import { Image, ComponentClassNames } from '@aws-amplify/ui-react';
 import { useStorageURL } from '@aws-amplify/ui-react/internal';
@@ -6,24 +7,37 @@ import { useStorageURL } from '@aws-amplify/ui-react/internal';
 import type { StorageImageProps } from './types';
 
 export const StorageImage = ({
-  imgKey,
   accessLevel,
-  identityId,
+  className,
   fallbackSrc,
-  onStorageError,
+  identityId,
+  imgKey,
+  onStorageGetError,
   ...rest
 }: StorageImageProps): JSX.Element => {
-  const { url, error } = useStorageURL(
-    imgKey,
-    { level: accessLevel, identityId },
-    fallbackSrc
+  const options = React.useMemo(
+    () => ({
+      accessLevel,
+      identityId,
+    }),
+    [accessLevel, identityId]
   );
 
-  if (error && onStorageError) {
-    onStorageError(error);
-  }
+  const errorConfig = React.useMemo(
+    () => ({
+      fallbackURL: fallbackSrc,
+      onStorageGetError,
+    }),
+    [fallbackSrc, onStorageGetError]
+  );
+
+  const url = useStorageURL(imgKey, options, errorConfig);
 
   return (
-    <Image className={ComponentClassNames.StorageImage} src={url} {...rest} />
+    <Image
+      {...rest}
+      className={classNames(ComponentClassNames.StorageImage, className)}
+      src={url}
+    />
   );
 };