+++ packages/react/src/primitives/types/switchField.ts
@@ -6,6 +6,7 @@ import { BaseInputProps } from './input';
 import { BaseFieldProps, LabelPositions } from './field';
 import { ElementType, PrimitiveProps } from './view';
 
+/** @deprecated For internal use only */
 export interface BaseSwitchFieldProps
   extends BaseInputProps,
     BaseFieldProps,
@@ -90,5 +91,5 @@ export interface BaseSwitchFieldProps
   value?: string | number;
 }
 
-export type SwitchFieldProps<Element extends ElementType = 'input'> =
+export type SwitchFieldProps<Element extends ElementType = 'div'> =
   PrimitiveProps<BaseSwitchFieldProps, Element>;